+++ plone/app/event/tests/test_doctests.py
@@ -8,6 +8,7 @@ DOCFILES = [
     'webdav.txt'
 ]
 DOCMODS = [
+    'plone.app.event.base',
 ]
 
 def test_suite():