+++ plone/app/event/tests/test_occurrence.py
@@ -1,12 +1,19 @@
 import datetime
 import transaction
 import unittest2 as unittest
-from DateTime import DateTime
-from plone.event.interfaces import IEventAccessor
+import zope.component
+from plone.event.utils import pydt
 from plone.event.utils import tzdel
+from plone.registry.interfaces import IRegistry
 from plone.testing.z2 import Browser
-
+from zope.publisher.browser import TestRequest
+from zope.publisher.interfaces.browser import IBrowserView
 from plone.app.event.at.content import IATEvent
+from plone.app.event.base import get_occurrences
+from plone.app.event.base import get_portal_events
+from plone.app.event.base import localized_now
+from plone.app.event.interfaces import IEventAccessor
+from plone.app.event.interfaces import IEventSettings
 from plone.app.event.interfaces import IOccurrence
 from plone.app.event.occurrence import Occurrence
 from plone.app.event.occurrence import OccurrenceTraverser
@@ -14,18 +21,20 @@ from plone.app.event.testing import PAEventAT_INTEGRATION_TESTING
 from plone.app.testing import TEST_USER_ID, TEST_USER_PASSWORD
 from plone.app.testing import setRoles
 
+
 class TestTraversal(unittest.TestCase):
 
     layer = PAEventAT_INTEGRATION_TESTING
 
     def setUp(self):
         self.portal = self.layer['portal']
+        reg = zope.component.getUtility(IRegistry)
+        settings = reg.forInterface(IEventSettings, prefix="plone.app.event")
+        settings.portal_timezone = "Australia/Brisbane"
+
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
         self.portal.invokeFactory(type_name='Event', id='at',
-                                  title='Event1',
-                                  timezone="Australia/Brisbane",
-                                  start=DateTime('Australia/Brisbane'),
-                                  end=DateTime('Australia/Brisbane') + 1)
+                                  title='Event1')
         self.at = self.portal['at']
         self.at_traverser = OccurrenceTraverser(self.at, self.layer['request'])
 
@@ -34,6 +43,11 @@ class TestTraversal(unittest.TestCase):
             AttributeError,
             self.at_traverser.publishTraverse,
             self.layer['request'], 'foo')
+
+    def test_default_views(self):
+        view = self.at_traverser.publishTraverse(
+            self.layer['request'], 'event_view')
+        self.assertTrue(IBrowserView.providedBy(view))
 
     def test_occurrence(self):
         self.at.setRecurrence('RRULE:FREQ=WEEKLY;COUNT=10')
@@ -45,9 +59,9 @@ class TestTraversal(unittest.TestCase):
             self.at_traverser.publishTraverse,
             self.layer['request'], str(qdate))
 
-        qdate = datetime.date.today() + datetime.timedelta(days=7)
+        qdatedt = pydt(self.at.start() + 7)
         item = self.at_traverser.publishTraverse(self.layer['request'],
-                                                 str(qdate))
+                                                 str(qdatedt.date()))
         self.assertTrue(IOccurrence.providedBy(item))
         self.assertTrue(IATEvent.providedBy(item.aq_parent))
 
@@ -62,24 +76,15 @@ class TestTraversal(unittest.TestCase):
         self.assertEqual(start, data['start'])
 
 
-class TestTraversalBrowser(unittest.TestCase):
-
-    layer = PAEventAT_INTEGRATION_TESTING
-
-    def setUp(self):
-        self.portal = self.layer['portal']
-        setRoles(self.portal, TEST_USER_ID, ['Manager'])
-        self.portal.invokeFactory(type_name='Event', id='at',
-                                  title='Event1',
-                                  timezone="Australia/Brisbane",
-                                  start=DateTime('Australia/Brisbane'),
-                                  end=DateTime('Australia/Brisbane') + 1)
-        self.portal['at'].setRecurrence('RRULE:FREQ=WEEKLY;COUNT=10')
-        transaction.commit()
+class TestTraversalBrowser(TestTraversal):
 
     def test_traverse_occurrence(self):
-        qdate = datetime.date.today() + datetime.timedelta(days=7)
-        url = '/'.join([self.portal['at'].absolute_url(), str(qdate)])
+        self.at.setRecurrence('RRULE:FREQ=WEEKLY;COUNT=10')
+        transaction.commit()
+
+        qdatedt = pydt(self.at.start() + 7)
+        url = '/'.join([self.portal['at'].absolute_url(),
+                        str(qdatedt.date())])
 
         browser = Browser(self.layer['app'])
         browser.addHeader('Authorization', 'Basic %s:%s' % (
@@ -87,6 +92,77 @@ class TestTraversalBrowser(unittest.TestCase):
         browser.open(url)
         self.assertTrue(
             self.portal['at'].title.encode('ascii') in browser.contents)
+
+
+class TestOccurrences(unittest.TestCase):
+
+    layer = PAEventAT_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        default_tz = 'CET'
+
+        reg = zope.component.getUtility(IRegistry)
+        settings = reg.forInterface(IEventSettings, prefix="plone.app.event")
+        settings.portal_timezone = default_tz
+
+        now = localized_now()
+        yesterday = now - datetime.timedelta(days=1)
+
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.portal.invokeFactory(
+            'Event',
+            'daily',
+            title=u'Daily Event',
+            start=now,
+            end=now + datetime.timedelta(hours=1),
+            location=u'Vienna',
+            recurrence='RRULE:FREQ=DAILY;COUNT=4',
+            timezone=default_tz,
+            whole_day=False)
+
+        self.portal.invokeFactory(
+            'Event',
+            'interval',
+            title=u'Interval Event',
+            start=yesterday,
+            end=yesterday + datetime.timedelta(hours=1),
+            location=u'Halle',
+            recurrence='RRULE:FREQ=DAILY;INTERVAL=2;COUNT=5',
+            timezone=default_tz,
+            whole_day=False)
+
+        self.now = now
+        self.yesterday = yesterday
+        self.daily = self.portal['daily']
+        self.interval = self.portal['interval']
+
+    def test_get_occurrences(self):
+        result = get_occurrences(self.portal,
+                                 get_portal_events(self.portal))
+        self.assertTrue(len(result) == 9)
+
+        result = get_occurrences(self.portal,
+                                 get_portal_events(self.portal), limit=5)
+        self.assertTrue(len(result) == 5)
+        self.assertTrue(IOccurrence.providedBy(result[0]))
+
+    def test_view_get_data_startdate(self):
+        future = localized_now() + datetime.timedelta(days=5)
+        form = dict(start=str(future.date()))
+        request = TestRequest(form=form)
+        view = zope.component.getMultiAdapter(
+            (self.portal, request), name='occurrences.html')
+
+        result = view.get_data()
+        self.assertEqual(2, len(result))
+
+    def test_view_get_data_invalid(self):
+        request = TestRequest(form=dict(start='invalid'))
+        view = zope.component.getMultiAdapter(
+            (self.portal, request), name='occurrences.html')
+        result = view.get_data()
+        self.assertEqual(9, len(result))
 
 
 def test_suite():